% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl2sz.R
\name{dl2sz}
\alias{dl2sz}
\title{przekształca dane z postaci długiej do szerokiej}
\usage{
dl2sz(dane, klucz, id = character(), zmienne = character())
}
\arguments{
\item{dane}{ramka danych z danymi w postaci "pośredniej"}

\item{klucz}{kolumna zawierająca sufiksy, które stworzą nowe kolumny}

\item{id}{wektor nazw zmiennych będących kluczami (dane wyjściowe będą miały
tyle wierszy, ile jest unikalnych kombinacji wartości tych kolumn w danych
wejściowych)}

\item{zmienne}{wektor nazw zmiennych, które mają zostać przetworzone do
postaci "szerokiej"}
}
\description{
\code{\link[reshape2]{dcast}} oraz \code{\link[tidyr]{spread}} 
  są o tyle niewygodne, że wymagają danych w postaci "kanonicznie długiej". 
  Tymczasem wygodnie jest móc przekształcić do postaci "prawdziwie szerokiej"
  dane w formie "pośredniej", np. `okres` (wartości 1, 2, 3), `zm1`, `zm2` 
  przekształcić na `zm1_1`, `zm1_2`, `zm1_3`, `zm2_1`, `zm2_2`, `zm2_3`.
  
  Funkcja dostarcza skrótowej składni do wykonania takiej właśnie operacji.
  
  Wewnętrznie wykonuje \code{\link[tidyr]{spread}} na każdej z wyznaczonych 
  zmiennych, a następnie łączy wyniki za pomocą
  \code{\link[dplyr]{full_join}}.
}
\details{
Jeśli parametr \code{id} (lub \code{zmienne}) nie zostanie podany,
zostanie on wywiedziony jako "wszystkie kolumny danych wejściowych z
wyłączeniem tych, na które wskazuje parametr \code{klucz} oraz \code{zmienne}
(\code{id})". Oznacza to oczywiście, że przynajmnie jeden z parametrów
\code{id} i \code{zmienne} musi zostać podany.
}
