% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/konfiguruj.R
\name{.konfiguruj}
\alias{.konfiguruj}
\title{Konfiguruje działanie pakietu}
\usage{
.konfiguruj(giodo = 3)
}
\arguments{
\item{giodo}{minimalna liczba obserwacji w grupie (statystyki dla mniejszych
grup są anonimizowane)}
}
\value{
wektor lub ramka danych
}
\description{
Konfiguruje działanie pakietu
}
