% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etykietyHistogramu.R
\name{etykietyHistogramu}
\alias{etykietyHistogramu}
\title{Usuwa nachodzące na siebie etykiety osi}
\usage{
etykietyHistogramu(wykres, breaks, rozmiarTekstu, pomijajEtykiety)
}
\arguments{
\item{wykres}{obiekt wykresu ggplot}

\item{breaks}{wektor pozycji (i jednocześnie etykiet) osi}

\item{rozmiarTekstu}{rozmiar tekstu na wykresie}

\item{pomijajEtykiety}{liczba określająca minimalną odległość między
wyświetlanymi etykietami}
}
\value{
zmodyfikowany obiekt wykresu ggplot
}
\description{
Usuwa nachodzące na siebie etykiety osi
}
