% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G.R
\name{G}
\alias{G}
\title{wyznacza grupę ze względu na kwantyle}
\usage{
G(x, q, n, filtr = NULL)
}
\arguments{
\item{x}{wartości wg których nastąpi podział na grupy}

\item{q}{numer kwantyle}

\item{n}{liczba kwantyli}

\item{filtr}{globalny filtr obserwacji}
}
\value{
[logical.vector]
}
\description{
Zwraca wektor wybierający obserwacje, które jednocześnie:
\itemize{
  \item należą do grupy wyznaczanej przez parametr filtr
  \item w ramach w. w. grupy należą do kwantyla q (spośród
    n kwantyli) ze względu na wartość zmiennej przekazanej
    w parametrze x
}
}

