% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generujRaporty.R
\name{generujRaporty}
\alias{generujRaporty}
\title{generuje raporty na podstawie szablonu danych oraz opisu grup
  odbiorców}
\usage{
generujRaporty(plikSzablonu, grupyOdbiorcow, dane,
  daneMiesieczne = data.frame(), katalogWy = "", prefiksPlikow = "",
  ramkiTablic = FALSE, sprzataj = TRUE, kontynuujPoBledzie = TRUE)
}
\arguments{
\item{plikSzablonu}{ścieżka do pliku Markdown z szablonem raportu}

\item{grupyOdbiorcow}{ramka danych (lub lista), której kolejne elementy
określają grupy odbiorców}

\item{dane}{ramka danych lub ścieżka do pliku z danymi}

\item{daneMiesieczne}{ramka danych lub ścieżka do pliku z danymi miesięcznymi}

\item{katalogWy}{katalog, w którym zapisane zostaną wygenerowane raporty 
(względem pliku szablonu); uwaga! jeśli podany, wtedy pełna ścieżka 
katalogu wyjściowego nie może zawierać polskich znaków ani spacji (sic!)}

\item{prefiksPlikow}{prefiks nazw plików wygenerowanych raportów}

\item{ramkiTablic}{czy przekształcać wyjściowy kod TeX-a w celu uzyskania
ramek w tabelach}

\item{sprzataj}{czy usuwać pliki tymczasowe wytworzone w trakcie generowania
ostatecznych PDF-ów}

\item{kontynuujPoBledzie}{czy kontynuować generowania raportów jeśli podczas
generowania jednego z nich wystąpił błąd}
}
\description{
Pliki raportów tworzone są ze złączenia prefiksu z wartością z
pierwszej kolumny ramki danych opisującej odbiorców (a gdy odbiorcy opisywani
są listą przez nazwy elementów listy).

Kolumny ramki danych parametru 'grupyOdbiorcow' odpowiadają zmiennym używanym
w szablonie Markdown raportu, a wiersze kolejnym grupom odbiorców. Przy tym 
te zmienne ramki danych, których nazwy zaczynają się kropką są: \itemize{ 
\item ewaluowane w środowisku danych (a więc w ten sposób należy definiować
wszelkie filtry) \item przed przekazaniem do szablonu ich nazwa pozbawiana
jest wiodącej kropki }

Parametr 'grupy Odbiorcow' może być również listą. W taki wypadku 
przetwarzanie grup odbiorców polega na prostym ładowaniu zawartości kolejnych
elementów listy 'grupyOdbiorcow' jako zmiennych dostępnych w środowisku, a 
następnie kompilacji pliku szablonu raportu.
}

