% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Me}
\alias{Me}
\title{mediana}
\usage{
Me(x, wyrownaj = T, dokl = 2)
}
\arguments{
\item{x}{wektor wartości}

\item{wyrownaj}{czy wyrównywać długość wyniku (TRUE w wypadku osadzania w tabelach)}

\item{dokl}{liczba cyfr po przecinku, do których zaokrąglony zostanie wynik}
}
\description{
mediana
}

