% Generated by roxygen2 (4.0.2): do not edit by hand
\name{N}
\alias{N}
\title{liczba rekordów niebędących brakami danych}
\usage{
N(x, w = NULL, wyrownaj = T)
}
\arguments{
\item{x}{wektor wartości}

\item{w}{[opcjonalny] wektor zliczanych wartości}

\item{wyrownaj}{czy wyrównywać długość wyniku (TRUE w wypadku osadzania w
tabelach)}
}
\description{
Zlicza liczbę rekordów niebędących brakami danych.

Jeśli podano wektor w, zliczane są tylko wartości zawarte w wektorze w.
}

