% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/naLiczbe.R
\name{naLiczbe}
\alias{naLiczbe}
\title{Konwertuje przekazany wektor na wartości liczbowe}
\usage{
naLiczbe(dane)
}
\arguments{
\item{dane}{wektor do konwersji}
}
\value{
[numeric] skonwertowane wartości liczbowe
}
\description{
\code{as.numeric()}, który dodatkowo zachowuje nazwy elementów i pozbywa się
ew. kończącego znaku procenta
}

