% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/naNaWartosc.R
\name{naNaWartosc}
\alias{naNaWartosc}
\title{znajduje sensowną wartość (liczbę lub ciąg znaków), którą można
  zastąpić brak danych}
\usage{
naNaWartosc(x)
}
\arguments{
\item{x}{wektor wartości}
}
\value{
stosowny do typu x
}
\description{
znajduje sensowną wartość (liczbę lub ciąg znaków), którą można
  zastąpić brak danych
}

