% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obliczDaneMiesieczne.R
\name{obliczDaneMiesieczne}
\alias{obliczDaneMiesieczne}
\title{Agreguje dane do wykresów miesięcznych}
\usage{
obliczDaneMiesieczne(daneAbs, daneMies, zmienne, grupy = character(),
  dataRel = character(), od = character(), do = character(),
  odRel = numeric(), doRel = numeric(), statystyka = "mean")
}
\arguments{
\item{daneAbs}{ramka danych ze zmiennymi na poziomie absolwenta na kierunku
studiów}

\item{daneMies}{ramka danych ze zmiennymi na poziomie absolwenta na kierunku
studiów w posczególnych miesiącach}

\item{zmienne}{wektor nazw zmiennych, które będą agregowane (zmienne muszą
się znajdować albo w ramce danych \code{daneAba} albo w \code{daneMies})}

\item{grupy}{wektor nazw zmiennych grupujących (zmienne muszą
się znajdować albo w ramce danych \code{daneAba} albo w \code{daneMies})}

\item{dataRel}{nazwa zmiennej względem której liczony będzie czas (typowo
\code{DATA_ZAK} lub \code{DATA_ROZP} przekazywane w ramce danych
\code{daneAbs}); jeśli nie zostanie podana, czas liczony będzie w
miesiącach kalendarzowych}

\item{od}{minimalna wartość zmiennej \code{OKRES} w ramce danych
\code{daneAbs}, aby rekord został uwzględniony w analizie}

\item{do}{maksymalna wartość zmiennej \code{OKRES} w ramce danych
\code{daneAbs}, aby rekord został uwzględniony w analizie}

\item{odRel}{minimalna wartość relatywnie wyliczonego okresu, aby rekord
został uwzględniony w analizach}

\item{doRel}{maksymalna wartość relatywnie wyliczonego okresu, aby rekord
został uwzględniony w analizach}

\item{statystyka}{nazwa funkcji agregującej dane (domyślnie średnia); funkcja
musi obsługiwać parametr \code{na.rm = TRUE}}
}
\value{
[data.frame] ramka danych z obliczonymi wartościami
}
\description{
Funkcja wymaga jawnie przekazanych danych w odpowiednim formacie (ramki
danych \code{daneAbs} i \code{daneMies} muszą być łączliwe (zawierać zmienną
o takiej samej nazwie, typowo \code{ID_ZDAU}), a ramka danych \code{daneMies}
musi zawierać zmienną \code{OKRES}).
}

