% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/P.R
\name{P}
\alias{P}
\title{odsetek rekordów o zadanych wartościach}
\usage{
P(x, w, wyrownaj = T, dokl = 1)
}
\arguments{
\item{x}{wektor wartości}

\item{w}{wektor zliczanych wartości}

\item{wyrownaj}{czy wyrównywać długość wyniku (TRUE w wypadku osadzania w
tabelach)}

\item{dokl}{liczba cyfr po przecinku, do których zaokrąglony zostanie wynik}
}
\description{
Oblicza odsetek rekordów wektora x o wartościah zawartych w wektorze w.
}

