% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/polamTekst.R
\name{polamTekst}
\alias{polamTekst}
\title{Łamie przekazany wektor znaków}
\usage{
polamTekst(wektor, n = 15, wymus = FALSE)
}
\arguments{
\item{wektor}{wektor znaków do połamania}

\item{n}{maksymalna liczba znaków w linii}

\item{wymus}{czy łamać tekst, jeśli zawiera znaki nowej linii}
}
\value{
[character] połamany wektor znaków
}
\description{
Łamie przekazany wektor znaków, tak by długość pojedynczej linii nie
przekraczała zadanej liczby znaków.
}

