% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Q}
\alias{Q}
\title{wskazany kwantyl}
\usage{
Q(x, q, n, wyrownaj = T, dokl = 2)
}
\arguments{
\item{x}{wektor wartości}

\item{q}{numer kwantyla (może być wektorem)}

\item{n}{liczba kwantyli}

\item{wyrownaj}{czy wyrównywać długość wyniku (TRUE w wypadku osadzania w tabelach)}

\item{dokl}{liczba cyfr po przecinku, do których zaokrąglony zostanie wynik}
}
\description{
Istnieje możliwość obliczenia wartości wielu kwantyli jednocześnie -
wystarczy jako q przekazać wektor. W takim wypadku zwrócony zostanie wektor
odpowiednich kwantyli.
}

