% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{R}
\alias{R}
\title{współczynnik korelacji liniowej}
\usage{
R(x, y, wyrownaj = NA, dokl = 2)
}
\arguments{
\item{x}{wektor wartości pierwszej zmiennej}

\item{y}{wektor wartości drugiej zmiennej}

\item{wyrownaj}{czy wyrównywać długość wyniku (jeśli NA, wybór zostanie
dokonany automatycznie)}

\item{dokl}{liczba cyfr po przecinku, do których zaokrąglony zostanie wynik}
}
\description{
Oblicza współczynnik korelacji liniowej
}

