% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/R2.R
\name{R2}
\alias{R2}
\title{kwadrat współczynnika korelacji liniowej}
\usage{
R2(x, y, wyrownaj = NA, dokl = 2)
}
\arguments{
\item{x}{wektor wartości pierwszej zmiennej}

\item{y}{wektor wartości drugiej zmiennej}

\item{wyrownaj}{czy wyrównywać długość wyniku (jeśli NA, wybór zostanie
dokonany automatycznie)}

\item{dokl}{liczba cyfr po przecinku, do których zaokrąglony zostanie wynik}
}
\description{
kwadrat współczynnika korelacji liniowej
}

