% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/statKorelacja.R
\name{statKorelacja}
\alias{statKorelacja}
\title{oblicza dowolną z udostępnianych przez pakiet korelacji}
\usage{
statKorelacja(x, y, wyrownaj, dokl, metoda, kwadrat, call)
}
\arguments{
\item{x}{wektor wartości pierwszej zmiennej}

\item{y}{wektor wartości drugiej zmiennej}

\item{wyrownaj}{czy wyrównywać długość wyniku (TRUE w wypadku osadzania w
tabelach)}

\item{dokl}{liczba cyfr po przecinku, do których zaokrąglony zostanie wynik}

\item{metoda}{sposób liczenia korelacji - argument method funkcji cor()}

\item{kwadrat}{czy podnieść wynik do kwadratu?}

\item{call}{wywołanie funkcji (np. z sys.call()), do którego wyrównywany jest
wynik}
}
\description{
Wewnętrzna funkcja, która dba o jednakowy algorytm obliczania wszystkich
korelacji:
\itemize{
  \item sprawdza czy x i y są odpowiednimi wektorami (jeśli to możliwe,
    dokonuje potrzebnej konwersji)
  \item wywołuje na x i y funkcję giodo()
  \item usuwa z {x, y} wiersze z brakami danych
  \item jeśli {x, y} nie zawiera wierszy przyjmuje '-' jako wynik
  \item jeśli {x, y} zawiera wiersze, to liczy stosowną korelację i zaokrągla
    wynik zgodnie z parametrem dokl
  \item ew. wyrównuje długość wyniku do długości wywołania call
}
}

