% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/statWektor.R
\name{statWektor}
\alias{statWektor}
\title{oblicza wskazaną statystykę dla wskazanego wektora danych}
\usage{
statWektor(x, f, call, wyrownaj = T, dokl = 2)
}
\arguments{
\item{x}{wektor wartości lub ramka danych wartości}

\item{f}{funkcja do wykonania na x}

\item{call}{wywołanie, do którego długości wyrównywany jest wynik}

\item{wyrownaj}{czy wyrównywać długość wyniku (TRUE w wypadku osadzania w tabelach)}

\item{dokl}{liczba cyfr po przecinku, do których zaokrąglony zostanie wynik}
}
\value{
zależnie od funkcji f
}
\description{
\itemize{
  \item sprawdza czy x jest wektorem
  \item wywołuje na x funkcję giodo()
  \item usuwa z x braki danych
  \item jeśli wektor x jest pusty przyjmuje '-' jako wynik
  \item jeśli wektor x nie jest pusty, wywołuje na x funkcję f i
    zaokrągla wynik zgodnie z parametrem dokl
  \item ew. wyrównuje długość wyniku do długości wywołania call
}
}

