% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TAB.R
\name{TAB}
\alias{TAB}
\title{rysuje tablicę na podstawie ramki danych}
\usage{
TAB(dane)
}
\arguments{
\item{dane}{ramka danych do narysowania w postaci tablicy}
}
\value{
character vector
}
\description{
W wypadku, gdy liczba wierszy w tablicy jest nieprzewidywalna, nie da się opisać tablicy w szablonie raportu za pomocą kodu markdown.
Funkcja \code{TAB} pozwala obejść to ograniczenie rysując tablicę na podstawie dowolnej ramki danych.

Tablica rysowana jest w składni multiline tables, a więc z zapewnieniem łamania wierszy w komórkach.

Kolumny zawierające wartości liczbowe są wyrównane do prawej, zaś kolumny tekstowe do lewej (być może w przyszłości pojawi się możliwość ręcznego określania wyrównania).
}

