% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Tau.R
\name{Tau}
\alias{Tau}
\title{współczynnik korelacji rangowej Kendalla}
\usage{
Tau(x, y, wyrownaj = T, dokl = 2)
}
\arguments{
\item{x}{wektor wartości pierwszej zmiennej}

\item{y}{wektor wartości drugiej zmiennej}

\item{wyrownaj}{czy wyrównywać długość wyniku (TRUE w wypadku osadzania w tabelach)}

\item{dokl}{liczba cyfr po przecinku, do których zaokrąglony zostanie wynik}
}
\description{
współczynnik korelacji rangowej Kendalla
}

