% Generated by roxygen2 (4.0.2): do not edit by hand
\name{W}
\alias{W}
\title{wyrównanie długości}
\usage{
W(x)
}
\arguments{
\item{x}{wartość}
}
\description{
Zwraca przekazany argument wyrównując go spacjami z lewej, tak by długość
zwróconej wartości odpowiadała długości wstawki R.

Przydatna do wstawianie wartości do multiline tables w wypadku, gdy wyniki
obliczono wcześniej w innej wstawce R.
}

