% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wczytajCSV}
\alias{wczytajCSV}
\title{nakładka na read.csv2 ustawiająca kilka przydatnych wartości domyślnych}
\usage{
wczytajCSV(plik, ...)
}
\arguments{
\item{plik}{ścieżka do pliku CSV}

\item{...}{wszelkie argumenty funkcji read.csv2}
}
\description{
Wywołuje read.csv2() z parametrami stringsAsFactors = FALSE oraz fileEncoding = 'Windows-1250'
}

