% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wczytajDane.R
\name{wczytajDane}
\alias{wczytajDane}
\title{wczytuje dane z pliku rozpoznając typ pliku}
\usage{
wczytajDane(plik, kodowanie = "Windows-1250", separator = ";")
}
\arguments{
\item{plik}{ścieżka do pliku danych}

\item{kodowanie}{kodowanie kodowanie polskich znaków stosowane w plikach opisu grup i
danych (istotne tylko w wypadku plików CSV)}

\item{separator}{separator plików CSV (istotny tylko w wypadku plików CSV)}
}
\value{
wczytane dane
}
\description{
Wczytuje dane ze wskazanego pliku automatycznie dostosowując sposób ich 
wczytania do typu pliku (obecnie obsługiwane CSV, XLSX i RData).

W wypadku plików RData zwracane są wszystkie ramki danych zawarte w pliku.
Jeśli jest ich wiele, zwracana jest ich lista, jeśli tylko jedna, zwracana
jest wprost jako ramka danych.
}

