% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wczytajOdbiorce}
\alias{wczytajOdbiorce}
\title{wczytuje wskazanego odbiorcę}
\usage{
wczytajOdbiorce(grupy, dane = data.frame(), n = 1)
}
\arguments{
\item{grupy}{ramka danych, lista lub ścieżka do pliku CSV z definicjami grup
odbiorców}

\item{dane}{ramka danych lub ścieżka do pliku CSV z danymi}

\item{n}{numer odbiorcy do wczytania}
}
\value{
[list] definicja odbiorcy
}
\description{
Wczytuje z ramki danych, listy lub pliku CSV z definicjami odbiorców definicję
wskazanego odbiorcy (domyślnie pierwszego).

Jeśli definicja odbiorcy zawiera zmienne, które powinny zostać zewaluowane
w kontekście danych (ich nazwa zaczyna się od kropki), dokonuje ich ewaluacji
(i usuwa w. w. kropkę z nazw zmiennych).

Do definicji odbiorcy dołączane są również przekazane dane.
}
\details{
Jeśli w środowisku ustawiona jest zmienna .nieWczytujOdbiorcy, wtedy funkcja
zwraca pustą listę. Służy to pominięciu wywołania wczytajOdbiorce() w treści
szablonu w momencie wsadowego generowania raportów, np. funkcją
generujRaporty() (którą ustawia zmienną .nieWczytujOdbiorcy).
}

