% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wstawTresc.R
\name{wstawTresc}
\alias{wstawTresc}
\title{wykonuje wstawki R w danym kodzie Markdown i wypisuje wynik za pomocą
  cat}
\usage{
wstawTresc(kod)
}
\arguments{
\item{kod}{kod Markdown (z ew. wstawkami R) do wykonania lub ścieżka do pliku .Rmd}
}
\value{
przetworzony kod
}
\description{
wykonuje wstawki R w danym kodzie Markdown i wypisuje wynik za pomocą
  cat
}

