% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wykresDefaultTheme}
\alias{wykresDefaultTheme}
\title{Domyślny styl wykresów}
\usage{
wykresDefaultTheme(wykres, rozmiarTekstu)
}
\arguments{
\item{wykres}{obiekt gpplot2 wykresu}

\item{rozmiarTekstu}{domyślny rozmiar tekstu}
}
\description{
Funkcja zwracająca domyślne ostylowania wykresów, współdzielone pomiędzy
wszystkie funkcje rysujące wykresy
}

