% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wykresHistogram}
\alias{wykresHistogram}
\title{Wykres - histogram}
\usage{
wykresHistogram(dane, n = 30, rozmiarTekstu = 16, opcjeWykresu = NULL)
}
\arguments{
\item{dane}{wektor danych}

\item{n}{liczba przedziałów}

\item{rozmiarTekstu}{bazowy rozmiar tekstu}

\item{opcjeWykresu}{dodatkowe opcje wykresu (zostaną dodane do obiektu wykresu ggplot2)}
}
\value{
[gg] obiekt wykresu pakietu ggplot2
}
\description{
Rysuje histogram z przekazanych danych.
}

