% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wykresKolowy}
\alias{wykresKolowy}
\title{Wykres kołowy}
\usage{
wykresKolowy(dane, etykiety = NULL, rozmiarTekstu = 16,
  opcjeWykresu = NULL)
}
\arguments{
\item{dane}{wektor danych}

\item{etykiety}{opcjonalny wektor z etykietami wartości}

\item{rozmiarTekstu}{bazowy rozmiar tekstu}

\item{opcjeWykresu}{dodatkowe opcje wykresu (zostaną dodane do obiektu wykresu ggplot2)}
}
\value{
[gg] obiekt wykresu pakietu ggplot2
}
\description{
Rysuje wykres kołowy.

Jeśli dane zawierają duplikaty wartości lub są factor-em, wtedy narysowane
zostaną częstości występowania poszczególnych wartości.
W przeciwnym wypadku narysowane zostaną przekazane dane.

Właściwe etykiety wartości uzyskać można w dwojaki sposób - albo przekazując
do funkcji factor (wektor z przypisanymi etykietami wartości) albo używając
argumentu "etykiety". Argument "etykiety" powinien mieć postać:

c('wartość1' = 'etykieta1', 'wartość2' = 'etykieta2', ...)
}

