% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykresKolowyNorm.R
\name{wykresKolowyNorm}
\alias{wykresKolowyNorm}
\title{Wykres kołowy z normalizacją przekazanych danych do 100 procent}
\usage{
wykresKolowyNorm(dane, dokl = 1, tytul = "", rozmiarTekstu = NULL,
  opcjeWykresu = NULL, rysuj = TRUE)
}
\arguments{
\item{dane}{wektor danych}

\item{dokl}{do ilu miejsc po przecinku zaokrąglić wartości}

\item{tytul}{tytuł wykresu}

\item{rozmiarTekstu}{bazowy rozmiar tekstu}

\item{opcjeWykresu}{dodatkowe opcje wykresu (zostaną dodane do obiektu wykresu ggplot2)}

\item{rysuj}{czy funkcja ma narysować wykres czy tylko zwrócić wygenerowany obiekt wykresu}
}
\value{
[gg] obiekt wykresu pakietu ggplot2
}
\description{
Rysuje wykres kołowy.

Jeśli dane nie sumują się do 100, dokonuje stosownej normalizacji.
}

