% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykresKolowyZlicz.R
\name{wykresKolowyZlicz}
\alias{wykresKolowyZlicz}
\title{Wykres kołowy ze zliczenia wartości zmiennej}
\usage{
wykresKolowyZlicz(dane, etykiety = NULL, tytul = "", rozmiarTekstu = NULL,
  opcjeWykresu = NULL, rysuj = TRUE)
}
\arguments{
\item{dane}{wektor danych}

\item{etykiety}{opcjonalny wektor z etykietami wartości}

\item{tytul}{tytuł wykresu}

\item{rozmiarTekstu}{bazowy rozmiar tekstu}

\item{opcjeWykresu}{dodatkowe opcje wykresu (zostaną dodane do obiektu wykresu ggplot2)}

\item{rysuj}{czy funkcja ma narysować wykres czy tylko zwrócić wygenerowany obiekt wykresu}
}
\value{
[gg] obiekt wykresu pakietu ggplot2
}
\description{
Rysuje wykres kołowy.

Odsetki oblicza jako częstości wystąpień poszczególnych wartości w 
przekazanym wektorze z pominięciem braków danych.

Właściwe etykiety wartości uzyskać można w dwojaki sposób - albo przekazując
do funkcji factor (wektor z przypisanymi etykietami wartości) albo używając
argumentu "etykiety". Argument "etykiety" powinien mieć postać:
\code{c('wartość1' = 'etykieta1', 'wartość2' = 'etykieta2', ...)}.

Jeśli argument "etykiety" zostanie użyty, ale nie będzie zawierał wszystkich
wartości występujących w danych, wartości te zostaną pominięte i nie wejdą
do podstawy procentowania!

W wypadku, gdy wektor danych nie będzie factorem i nie zostanie podany 
argument "etykiety", wtedy za etykiety zostaną przyjęte poszczególne 
wartości.
}

