% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wykresPusty.R
\name{wykresPusty}
\alias{wykresPusty}
\title{Pusty wykres}
\usage{
wykresPusty(tytul = NULL, tytulX = NULL, tytulY = NULL, rysuj = TRUE)
}
\arguments{
\item{tytul}{tytuł wykresu}

\item{tytulX}{tytuł osi X wykresu}

\item{tytulY}{tytuł osi Y wykresu}

\item{rysuj}{czy funkcja ma narysować wykres czy tylko zwrócić wygenerowany obiekt wykresu}
}
\value{
[gg] obiekt wykresu pakietu ggplot2
}
\description{
Funkcja używana do obsługi błędów, np. przy próbie narysowania wykresu z
pustych lub niepoprawnych danych.
}

