% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wykresRozrzutu.R
\name{wykresRozrzutu}
\alias{wykresRozrzutu}
\title{Wykres rozrzutu}
\usage{
wykresRozrzutu(x, y, rozmiar = NULL, etykiety = NULL, tytul = "",
  tytulX = NULL, tytulY = NULL, rozmiarTekstu = NULL,
  opcjeWykresu = NULL, rysuj = TRUE)
}
\arguments{
\item{x}{wektor pozycji na osi X}

\item{y}{wektor pozycji na osi Y}

\item{rozmiar}{rozmiar poszczególnych punktów}

\item{etykiety}{etykiety punktów na wykresie}

\item{tytul}{tytuł wykresu}

\item{tytulX}{tytuł osi X wykresu}

\item{tytulY}{tytuł osi Y wykresu}

\item{rozmiarTekstu}{bazowy rozmiar tekstu}

\item{opcjeWykresu}{dodatkowe opcje wykresu (zostaną dodane do obiektu wykresu ggplot2)}

\item{rysuj}{czy funkcja ma narysować wykres czy tylko zwrócić wygenerowany obiekt wykresu}
}
\value{
[gg] obiekt wykresu pakietu ggplot2
}
\description{
Rysuje wykres rozrzutu (X-Y).
}

