% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykresRozrzutuIloraz.R
\name{wykresRozrzutuIloraz}
\alias{wykresRozrzutuIloraz}
\title{Wykres rozrzutu dla skal ilorazowych}
\usage{
wykresRozrzutuIloraz(x, y, etykiety = NULL, rozmiar = NULL, tytul = "",
  tytulX = NULL, tytulY = NULL, minX = 1, minY = 1, margX = 0.1,
  margY = 0.1, krokX = 0.25, krokY = 0.25, maxRozmPkt = 5,
  rozmiarMin = 10, rozmiarTekstu = NULL, opcjeWykresu = NULL,
  rysuj = TRUE)
}
\arguments{
\item{x}{wektor pozycji na osi X}

\item{y}{wektor pozycji na osi Y}

\item{etykiety}{etykiety punktów na wykresie}

\item{rozmiar}{rozmiar poszczególnych punktów}

\item{tytul}{tytuł wykresu}

\item{tytulX}{tytuł osi X wykresu}

\item{tytulY}{tytuł osi Y wykresu}

\item{minX}{minimalna wartość końca osi X}

\item{minY}{minimalna wartość końca osi Y}

\item{margX}{margines na etykiety punktów (obszar wykresu będzie większy o
margines od skrajnych wartości danych)}

\item{margY}{margines na etykiety punktów (obszar wykresu będzie większy o
margines od skrajnych wartości danych)}

\item{krokX}{krok etykiet na skali X}

\item{krokY}{krok etykiet na skali Y}

\item{maxRozmPkt}{rozmiar punktu na wykresie odpowiadający największej
wartości parametru \code{rozmiar}}

\item{rozmiarMin}{minimalna wartość rozmiaru danego punktu, aby został
zaprezentowany na wykresie (brana pod uwagę tylko gdy podano parametr
\code{rozmiar})}

\item{rozmiarTekstu}{bazowy rozmiar tekstu na wykresie}

\item{opcjeWykresu}{dodatkowe opcje wykresu (zostaną dodane do obiektu wykresu ggplot2)}

\item{rysuj}{czy funkcja ma narysować wykres czy tylko zwrócić wygenerowany obiekt wykresu}
}
\value{
[gg] obiekt wykresu pakietu ggplot2
}
\description{
Rysuje wykres rozrzutu (X-Y) zakładając, że na obydwu osiach znajdują się
ilorazy, a więc początek układu współrzędnych powinien wypadać w punkcie (0, 0).
}

