% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wykresSlupkowy.R
\name{wykresSlupkowy}
\alias{wykresSlupkowy}
\title{Wykres słupkowy}
\usage{
wykresSlupkowy(dane, skumulowany = F, tytul = "", tytulX = NULL,
  tytulY = NULL, sufiksY = "", rozmiarTekstu = NULL,
  opcjeWykresu = NULL)
}
\arguments{
\item{dane}{wektor danych}

\item{skumulowany}{czy wykres ma być skumulowany?}

\item{tytul}{tytuł wykresu}

\item{tytulX}{tytuł osi X wykresu}

\item{tytulY}{tytuł osi Y wykresu}

\item{sufiksY}{ew. ciąg znaków dołączany do etykiet osi Y (np. znak procenta)}

\item{rozmiarTekstu}{bazowy rozmiar tekstu}

\item{opcjeWykresu}{dodatkowe opcje wykresu (zostaną dodane do obiektu wykresu ggplot2)}
}
\value{
[gg] obiekt wykresu pakietu ggplot2
}
\description{
Rysuje wykres słupkowy.
}

