% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/E.R
\name{E}
\alias{E}
\title{średnia}
\usage{
E(x, wyrownaj = NA, dokl = 2)
}
\arguments{
\item{x}{wektor wartości}

\item{wyrownaj}{czy wyrównywać długość wyniku (jeśli NA, wybór zostanie
dokonany automatycznie)}

\item{dokl}{liczba cyfr po przecinku, do których zaokrąglony zostanie wynik}
}
\description{
średnia
}

