% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generujRaporty.R
\name{generujRaporty}
\alias{generujRaporty}
\title{generuje raporty na podstawie szablonu danych oraz opisu grup odbiorców}
\usage{
generujRaporty(plikSzablonu, dane, grupyOdbiorcow, katalogWy = "",
  prefiksPlikow = "")
}
\arguments{
\item{plikSzablonu}{ścieżka do pliku Markdown z szablonem raportu}

\item{dane}{ramka danych lub ścieżka do pliku CSV z danymi}

\item{grupyOdbiorcow}{ramka danych (lub lista), której kolejne elementy określają grupy odbiorców}

\item{katalogWy}{katalog, w którym zapisane zostaną wygenerowane raporty (względem pliku szablonu)}

\item{prefiksPlikow}{prefiks nazw plików wygenerowanych raportów}
}
\description{
Pliki raportów tworzone są ze złączenia prefiksu z wartością z pierwszej
kolumny ramki danych opisującej odbiorców (a gdy odbiorcy opisywani są listą
przez nazwy elementów listy).

Kolumny ramki danych parametru 'grupyOdbiorcow' odpowiadają zmiennym używanym
w szablonie Markdown raportu, a wiersze kolejnym grupom odbiorców. Przy tym
te zmienne ramki danych, których nazwy zaczynają się kropką są:
\itemize{
  \item ewaluowane w środowisku danych (a więc w ten sposób należy
    definiować wszelkie filtry)
  \item przed przekazaniem do szablonu ich nazwa pozbawiana jest wiodącej
    kropki
}

Parametr 'grupy Odbiorcow' może być również listą. W taki wypadku
przetwarzanie grup odbiorców polega na prostym ładowaniu zawartości kolejnych
elementów listy 'grupyOdbiorcow' jako zmiennych dostępnych w środowisku, a
następnie kompilacji pliku szablonu raportu.
}

