% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generujRaporty}
\alias{generujRaporty}
\title{generuje raporty na podstawie szablonu danych oraz opisu grup odbiorców}
\usage{
generujRaporty(plikSzablonu, dane, grupyOdbiorcow, katalogWy = "",
  prefiksPlikow = "")
}
\arguments{
\item{plikSzablonu}{ścieżka do pliku Markdown z szablonem raportu}

\item{dane}{ramka danych lub ścieżka do pliku CSV z danymi}

\item{grupyOdbiorcow}{lista, której kolejne elementy określają grupy odbiorców}

\item{katalogWy}{katalog, w którym zapisane zostaną wygenerowane raporty (względem pliku szablonu)}

\item{prefiksPlikow}{prefiks nazw plików wygenerowanych raportów}
}
\description{
Przetwarzanie grup odbiorców polega na prostym ładowaniu zawartości kolejnych
elementów listy 'grupyOdbiorcow' jako zmiennych dostępnych w środowisku,
a następnie kompilacji pliku szablonu raportu. Tak więc zawartość elementów
listy 'grupyOdbiorcow' determinowana jest przez zmienne używane w szablonie
Markdown raportu.

Aby wprowadzić tu pewne uporządkowanie najlepiej trzymać się wytycznych:
\itemize{
  \item filtr definiujący najszerszą populację badaną w raporcie nazywamy 'grupaGl'
  \item wszelkie stałe nazywamy 'stNazwaStałej'
}

Parametr 'grupyOdbiorcow' może być również ramką danych. W takim wypadku
kolumny ramki danych odpowiadają zmiennym używanym w szablonie Markdown
raportu, a wiersze kolejnym grupom odbiorców. Przy tym te zmienne ramki
danych, których nazwy zaczynają się kropką są:
\itemize{
  \item ewaluowane w środowisku danych (a więc w ten sposób należy
    definiować wszelkie filtry)
  \item przed przekazaniem do szablonu ich nazwa pozbawiana jest wiodącej
    kropki
}
}

