% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/giodo.R
\name{giodo}
\alias{giodo}
\title{weryfikuje czy dane spełniają kryterium minimalnej liczebności grupy}
\usage{
giodo(x)
}
\arguments{
\item{x}{wektor danych lub ramka danych}
}
\value{
wektor lub ramka danych
}
\description{
Jeśli przekazane dane nie spełniają warunku minimalnej liczebności grupy,
wtedy wszystkie wartości zostają zamienione na braki danych (zanonimizowane).

Przy tym jeśli dane są długości 0, wtedy nie są anonimizowane.
}

