% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Me.R
\name{Me}
\alias{Me}
\title{mediana}
\usage{
Me(x, wyrownaj = NA, dokl = 2)
}
\arguments{
\item{x}{wektor wartości}

\item{wyrownaj}{czy wyrównywać długość wyniku (jeśli NA, wybór zostanie
dokonany automatycznie)}

\item{dokl}{liczba cyfr po przecinku, do których zaokrąglony zostanie wynik}
}
\description{
Oblicza medianę
}

