% Generated by roxygen2 (4.0.2): do not edit by hand
\name{N}
\alias{N}
\title{liczba rekordów niebędących brakami danych}
\usage{
N(x, wyrownaj = T)
}
\arguments{
\item{x}{wektor wartości}

\item{wyrownaj}{czy wyrównywać długość wyniku (TRUE w wypadku osadzania w tabelach)}
}
\description{
liczba rekordów niebędących brakami danych
}

