% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naNaWartosc.R
\name{naNaWartosc}
\alias{naNaWartosc}
\title{zastępuje braki danych}
\usage{
naNaWartosc(x)
}
\arguments{
\item{x}{wektor wartości}
}
\value{
stosowny do typu x
}
\description{
Znajduje sensowną wartość (liczbę lub ciąg znaków), którą można zastąpić brak
danych
}

