% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pw.R
\name{Pw}
\alias{Pw}
\title{odsetek rekordów o zadanych wartościach z pominięciem braków danych}
\usage{
Pw(x, w = TRUE, wyrownaj = NA, dokl = 1, znakProcent = NA)
}
\arguments{
\item{x}{wektor wartości}

\item{w}{wektor zliczanych wartości}

\item{wyrownaj}{czy wyrównywać długość wyniku (jeśli NA, wybór zostanie
dokonany automatycznie)}

\item{dokl}{liczba cyfr po przecinku, do których zaokrąglony zostanie wynik}

\item{znakProcent}{czy dostawić do zwróconej wartości znak procenta}
}
\description{
Oblicza odsetek rekordów wektora \code{x} o wartościah zawartych w wektorze
\code{w}, przy czym braki danych są pomijane w podstawie procentowania.
}
