% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q.R
\name{Q}
\alias{Q}
\title{wskazany kwantyl}
\usage{
Q(x, q, n, wyrownaj = NA, dokl = 2)
}
\arguments{
\item{x}{wektor wartości}

\item{q}{numer kwantyla (może być wektorem)}

\item{n}{liczba kwantyli}

\item{wyrownaj}{czy wyrównywać długość wyniku (jeśli NA, wybór zostanie
dokonany automatycznie)}

\item{dokl}{liczba cyfr po przecinku, do których zaokrąglony zostanie wynik}
}
\description{
Istnieje możliwość obliczenia wartości wielu kwantyli jednocześnie - 
wystarczy jako q przekazać wektor. W takim wypadku zwrócony zostanie wektor
odpowiednich kwantyli.
}
