% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ustawWyrownaj.R
\name{ustawWyrownaj}
\alias{ustawWyrownaj}
\title{zgaduje, czy należy dokonać wyrównania długości zwracanej wartości}
\usage{
ustawWyrownaj(wyrownaj)
}
\arguments{
\item{wyrownaj}{obecna wartość wyrownaj}
}
\value{
logical
}
\description{
Jeśli parametr \code{wyrownaj} nie jest brakiem danych, jest po prostu 
zwracany. W wypadku, gdy jest brakiem danych, następuje sprawdzenie, czy 
środowisko znajdujące się o \code{offset} pozycji wyżej w hierarhii środowisk
jest tym samym, w którym załadowano pakiet MLAK. Jeśli tak, zwracana jest
wartość \code{TRUE}, jeśli nie, zwracana jest wartość \code{FALSE}.
Wyrównywanie zwracanej wartości ma bowiem sens w zasadzie jedynie w wypadku
zwracania jej wprost do wstawki markdown, jeśli natomiast będzie ona dalej
przetwarzana przez inne funkcje R, wtedy prawie na pewno nie powinna być
wyrównywana.
}
