% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wczytajCSV.R
\name{wczytajCSV}
\alias{wczytajCSV}
\title{wczytuje dane z CSV}
\usage{
wczytajCSV(
  plik,
  sep = ";",
  quote = "\\"",
  dec = ",",
  fileEncoding = "Windows-1250",
  header = TRUE,
  fill = T,
  comment.char = "",
  stringsAsFactors = FALSE,
  ...
)
}
\arguments{
\item{plik}{ścieżka do pliku CSV}

\item{sep}{separator kolumn (domyślnie ;)}

\item{quote}{separator tekstu (domyślnie ")}

\item{dec}{separator dziesiętny (domyślnie ,)}

\item{fileEncoding}{kodowanie pliku CSV (domyślnie Windows-1250)}

\item{header}{czy pierwszy wiersz to nagłówki kolumn (domyślnie TRUE)}

\item{fill}{czy wyrównywać NA linie krótsze od pozostałych}

\item{comment.char}{znak rozpoczynający linie będące komentarzem}

\item{stringsAsFactors}{czy przekształcać stringi na factor-y (domyślnie FALSE)}

\item{...}{wszelkie argumenty funkcji read.table}
}
\value{
[data.frame] wczytane dane
}
\description{
nakładka na read.table zapewniająca wartości domyślne zgodne z 
  formatem zapisu plików CSV przez Excel-a
}
