% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wczytajDane.R
\name{wczytajDane}
\alias{wczytajDane}
\title{wczytuje dane z pliku rozpoznając typ pliku}
\usage{
wczytajDane(plik)
}
\arguments{
\item{plik}{ścieżka do pliku danych}
}
\value{
wczytane dane
}
\description{
Wczytuje dane ze wskazanego pliku automatycznie dostosowując sposób ich
wczytania do typu pliku (obecnie obsługiwane CSV i RData).

W wypadku plików RData zwracane są wszystkie ramki danych zawarte w pliku.
Jeśli jest ich wiele, zwracana jest ich lista, jeśli tylko jedna, zwracana
jest wprost jako ramka danych.
}

