% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wczytajOdbiorce.R
\name{wczytajOdbiorce}
\alias{wczytajOdbiorce}
\title{wczytuje wskazanego odbiorcę}
\usage{
wczytajOdbiorce(grupy, dane = data.frame(), daneMiesieczne = data.frame(),
  n = 1, dolacz = TRUE, kodowanie = "Windows-1250", separator = ";")
}
\arguments{
\item{grupy}{ramka danych, lista lub ścieżka do pliku CSV z definicjami grup 
odbiorców}

\item{dane}{ramka danych lub ścieżka do pliku z danymi}

\item{daneMiesieczne}{ramka danych lub ścieżka do pliku z danymi miesięcznymi}

\item{n}{numer odbiorcy do wczytania}

\item{dolacz}{czy dołączyć (funkcją attach) wczytane dane do środowiska, w
którym funkcja została uruchomiona}

\item{kodowanie}{kodowanie polskich znaków stosowane w plikach opisu grup i
danych (istotne tylko w wypadku plików CSV)}

\item{separator}{separator plików CSV (istotny tylko w wypadku plików CSV)}
}
\value{
[list] definicja odbiorcy
}
\description{
Wczytuje z ramki danych, listy lub pliku CSV/XLSX z definicjami odbiorców
definicję
wskazanego odbiorcy (domyślnie pierwszego).

Jeśli definicja odbiorcy zawiera zmienne, które powinny zostać zewaluowane
w kontekście danych (ich nazwa zaczyna się od kropki), dokonuje ich ewaluacji
(i usuwa w. w. kropkę z nazw zmiennych).

Do definicji odbiorcy dołączane są również przekazane dane.
}
\details{
Funkcja sprawdza, czy, a jeśli tak, to gdzie, w ścieżce wywołań
znajduje się funkcja generujRaporty(). W wypadku, gdy generujRaporty()
znajduje się w ścieżce wywołań, ale wczytajOdbiorce() nie zostało wywołane
bezpośrednio z niej, zwracana jest pusta lista. Służy to pominięciu wywołania
wczytajOdbiorce() w treści szablonu w momencie wsadowego generowania raportów
funkcją generujRaporty().
}

