% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wstawTresc.R
\name{wstawTresc}
\alias{wstawTresc}
\title{ewaluuje i wstawia kod RMarkdown}
\usage{
wstawTresc(kod)
}
\arguments{
\item{kod}{kod Markdown (z ew. wstawkami R) do wykonania lub ścieżka do pliku .Rmd}
}
\value{
przetworzony kod
}
\description{
Wykonuje wstawki R w danym kodzie Markdown i wypisuje wynik za pomocą cat
}

