% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykresDefaultTheme.R
\name{wykresDefaultTheme}
\alias{wykresDefaultTheme}
\title{Domyślny styl wykresów}
\usage{
wykresDefaultTheme(wykres, tytul = NULL, tytulX = NULL, tytulY = NULL,
  rozmiarTekstu = NULL)
}
\arguments{
\item{wykres}{obiekt gpplot2 wykresu}

\item{tytul}{tytuł wykresu}

\item{tytulX}{tytuł osi X wykresu}

\item{tytulY}{tytuł osi Y wykresu}

\item{rozmiarTekstu}{domyślny rozmiar tekstu}
}
\description{
Funkcja zwracająca domyślne ostylowania wykresów, współdzielone pomiędzy
wszystkie funkcje rysujące wykresy
}

