% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wykresHistogram.R
\name{wykresHistogram}
\alias{wykresHistogram}
\title{Wykres - histogram}
\usage{
wykresHistogram(dane, n = 9, tytul = "", tytulX = NULL, tytulY = NULL,
  rozmiarTekstu = NULL, opcjeWykresu = NULL, rownePrzedzialy = FALSE,
  rysuj = TRUE)
}
\arguments{
\item{dane}{wektor danych}

\item{n}{liczba przedziałów}

\item{tytul}{tytuł wykresu}

\item{tytulX}{tytuł osi X wykresu}

\item{tytulY}{tytuł osi Y wykresu}

\item{rozmiarTekstu}{bazowy rozmiar tekstu}

\item{opcjeWykresu}{dodatkowe opcje wykresu (zostaną dodane do obiektu wykresu ggplot2)}

\item{rownePrzedzialy}{czy przedziały na osi X powinny być równe (TRUE) czy wyznaczane przez kwantyle (FALSE)}

\item{rysuj}{czy funkcja ma narysować wykres czy tylko zwrócić wygenerowany obiekt wykresu}
}
\value{
[gg] obiekt wykresu pakietu ggplot2
}
\description{
Rysuje histogram z przekazanych danych.
}

