% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wykresKolowy.R
\name{wykresKolowy}
\alias{wykresKolowy}
\title{Wykres kołowy}
\usage{
wykresKolowy(dane, tytul = "", rozmiarTekstu = NULL, opcjeWykresu = NULL,
  rysuj = TRUE)
}
\arguments{
\item{dane}{wektor danych}

\item{tytul}{tytuł wykresu}

\item{rozmiarTekstu}{bazowy rozmiar tekstu}

\item{opcjeWykresu}{dodatkowe opcje wykresu (zostaną dodane do obiektu wykresu ggplot2)}

\item{rysuj}{czy funkcja ma narysować wykres czy tylko zwrócić wygenerowany obiekt wykresu}
}
\value{
[gg] obiekt wykresu pakietu ggplot2
}
\description{
Rysuje wykres kołowy, tzn. stosunek każdej z przekazanych wartości do ich
sumy.

Uwaga - wartości wyświetlane w legendzie nie są standaryzowane, lecz
wyświetlane dokładnie w takiej postaci, w jakiej zostały przekazane. Aby
legenda zawierała wystandaryzowane odsetki, użyj funkcji wykresKolowyNorm().
}

