% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wyrownajDl.R
\name{wyrownajDl}
\alias{wyrownajDl}
\title{formatuje wartość do długości wywołania funkcji}
\usage{
wyrownajDl(wynik, call, dokl, delta = 4)
}
\arguments{
\item{wynik}{wartość do sformatowania}

\item{call}{wywołanie funkcji (np. z sys.call()), do którego wyrównywany jest wynik}

\item{dokl}{liczba cyfr po przecinku, do których zaokrąglony zostanie wynik}

\item{delta}{liczba znaków, która ma zostać doliczna do długości wywołania}
}
\value{
character
}
\description{
Dopełnia wartość spacjami po lewej stronie, tak by łączna długość
utworzonego łańcucha znaków była równa liczbie znaków kodu wywołania funkcji
przekazanej w parametrze "call" plus tyle znaków, ile wynosi parametr
"delta"
}

