% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wyrownajDl.R
\name{wyrownajDl}
\alias{wyrownajDl}
\title{Formatuje wartość do długości wywołania funkcji}
\usage{
wyrownajDl(wynik, call, dokl, delta = 4)
}
\arguments{
\item{wynik}{wartość do sformatowania}

\item{call}{wywołanie funkcji (np. z sys.call()), do którego wyrównywany jest wynik}

\item{dokl}{liczba cyfr po przecinku, do których zaokrąglony zostanie wynik}

\item{delta}{liczba znaków, która ma zostać doliczna do długości wywołania}
}
\value{
character
}
\description{
Dopełnia wartość spacjami po lewej stronie, tak by łączna długość
utworzonego łańcucha znaków była równa liczbie znaków kodu wywołania funkcji
przekazanej w parametrze "call" plus tyle znaków, ile wynosi parametr
"delta"

Jeśli w stosie wywołań odnaleziona zostanie funkcja mutate_, summarize_ bądź
summarise_, wtedy uznaje się, że wyrównywanie nie ma sensu (ze względu na to,
że wywołanie zawiera wtedy podstawione wartości argumentów i jest absurdalnie
długie, jak również praktycznie na pewno nie jest ono używane do zwracania
wartości wprost do raportu) i długość wyrównania jest automatycznie ustawiana
na \code{max(nchar(wynik))}. Niestety podczas stosowania summarize nie można
całkowicie zrezygnować z wyrównywania, bo w wypadku, gdy część grup zwróci
stringi, a część liczby, próbuje ono z uporem godnym lepszej sprawy rzutować
stringi na liczby, czego wynikiem są bardzo dziwne wartości liczbowe (z '-'
wychodzą np. bardzo, bardzo małe, ale bynajmniej nie zerowe wartości).
}

